Revert "Fix speaker scheduler diagnostics build"
This reverts commit b6cd14d41eadc836ff194437713dffd2153c5b50.

+++ DS4Windows/DS4Control/DualSenseBluetoothSpeakerPassthrough.cs
@@ -355,7 +355,7 @@ namespace DS4Windows
 
                 while (!stopping)
                 {
-                    RecordScheduleLateness(Stopwatch.GetTimestamp() - (long)nextFrame);
+                    RecordScheduleLateness(Stopwatch.GetTimestamp() - nextFrame);
                     Array.Clear(sourceFrame, 0, sourceFrame.Length);
                 Array.Clear(frame, 0, frame.Length);
                 int sourceFramesPerTick = GetSourceFramesPerTick();
